Moved tpc test classes to right directory
