+++ hazelcast/src/test/java/com/hazelcast/internal/tpc/PromiseTest.java
@@ -1,4 +1,4 @@
-package com.hazelcast.alto.engine;
+package com.hazelcast.internal.tpc;
 
 
 import com.hazelcast.internal.tpc.Fut;
