+++ hazelcast/src/test/java/com/hazelcast/internal/tpc/iobuffer/IOBufferTest.java
@@ -1,4 +1,4 @@
-package com.hazelcast.alto.engine.iobuffer;
+package com.hazelcast.internal.tpc.iobuffer;
 
 import com.hazelcast.internal.tpc.iobuffer.IOBuffer;
 import com.hazelcast.test.HazelcastParallelClassRunner;
