+++ hazelcast/src/test/java/com/hazelcast/internal/tpc/nio/NioAsyncSocket_IntegrationTest.java
@@ -1,4 +1,4 @@
-package com.hazelcast.alto.engine.nio;
+package com.hazelcast.internal.tpc.nio;
 
 import com.hazelcast.internal.tpc.nio.NioAsyncReadHandler;
 import com.hazelcast.internal.tpc.nio.NioAsyncServerSocket;
