+++ hazelcast/src/test/java/com/hazelcast/internal/tpc/nio/NioEventloop_Unsafe_Test.java
@@ -1,10 +1,9 @@
-package com.hazelcast.alto.engine.nio;
+package com.hazelcast.internal.tpc.nio;
 
-import com.hazelcast.internal.tpc.nio.NioEventloop;
+import com.hazelcast.internal.tpc.Eventloop_Unsafe_Test;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.QuickTest;
 import com.hazelcast.internal.tpc.Eventloop;
-import com.hazelcast.alto.engine.Eventloop_Unsafe_Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
